app/requirements.txt
PyYAML==6.0.1
retry==0.9.2
imcsdk==0.9.13
pyghmi==1.5.63
pyvim==3.0.3
pyvmomi==8.0.1.0.2
hidapi==0.14.0
pydantic==2.5.2
retry==0.9.2
redis==5.0.1
pymongo==4.6.1
Flask==3.0.0
Flask-Login==0.6.3       
Flask-Mail==0.9.1        
Flask-RESTful==0.3.10
flask-restx==1.3.0  
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.24
Werkzeug==3.0.1
paramiko==3.4.0
gunicorn==23.0.0 